requirements.txt
pandas==1.3.4
numpy==1.24.4
statsmodels== 0.14.0
scipy==1.5.0
matplotlib== 3.3.4
networkx==2.5.1
ete3==3.1.2
MIPMLP==1.2.1
samba-metric==0.0.8
